Move some functions to a tags.py module.
